Before implementing different loss methods.

+++ infrastructure/settings.py
@@ -5,7 +5,7 @@ import torch
 torch.set_printoptions(sci_mode=False, linewidth=400)
 
 DEVICE: str = "cuda:0"
-DTYPE: torch.dtype = torch.float64
+DTYPE: torch.dtype = torch.float32
 PROJECT_NAME: str = "KF_RNN"
 PROJECT_PATH: str = os.getcwd()[:os.getcwd().find(PROJECT_NAME)] + PROJECT_NAME
